requirements.txt
python-osc==1.8.3
PyYAML==6.0.2